test: assert the orderings from the YAML, not from the prose
The middle finding is the one that matters and it is correct: those
"assertions" grepped the concatenated Markdown for the COMMENTS that
explain each ordering. Reversing the actual actions while leaving the
comments intact passed. My mutation check was worthless for the same
reason — I mutated the comment text, which is the string the assertion
searched for.

Now the fenced blocks are extracted and compared by offset. Verified the
right way round: changing only the prose no longer fails anything, while
moving the disarm behind the give-up notification, dropping
continue_on_error, or putting the upper bound back into `conditions:`
each fail. The helper also rejects an ambiguous anchor, because the
first attempt matched "Giving up" inside a comment.

Also: a raising `valve.close_valve` aborted the sequence before the
terminal disarm, so past the cutoff the retry stayed unbounded —
continue_on_error, then decide. And the German enforcement moved to its
own file; the contract test had reached 434 lines.
